refactor(phase-1): lift shared helpers from main.rs into library modules
Moves 12 functions/types out of the binary's main.rs into proper library
modules so tune/mod.rs no longer reaches back into the binary namespace:

- app/loop_runtime.rs: MeasuredDt, LoopControls trait, apply_measured_dt,
  apply_runtime_interval, write_safe_cv, handle_dt_skip_state_write,
  emit_state_write_failure, flush_state_at_shutdown, open_log_optional,
  millis_round_u64
- app/adapters_build.rs: PvSourceConfig, CvSinkConfig enums (moved from
  cli/types.rs) plus build_pv_source / build_cv_sink builder functions
- app/socket_dispatch.rs (#[cfg(unix)]): SocketSideEffect, handle_socket_request,
  handle_socket_set, apply_gain_param — now takes &mut dyn LoopControls
- cli/error.rs: CliError struct extracted from main.rs

LoopArgs implements LoopControls; tune/mod.rs imports only from the library.
All 139 integration tests pass; one pre-existing failure in
state_fail_after_escalates_to_prominent_warning confirmed on unmodified main.
