+++ crates/pid-ctl/src/app/mod.rs
@@ -1,6 +1,10 @@
 //! Controller session scaffolding and persistence primitives for the application layer.
 
+pub mod adapters_build;
+pub mod loop_runtime;
 pub mod state_store;
+#[cfg(unix)]
+pub mod socket_dispatch;
 
 use crate::adapters::CvSink;
 use pid_ctl_core::{ConfigError, PidConfig, PidController, StepInput, StepResult};
